+++ changelog.d/698.changed.md
@@ -1,0 +1,1 @@
+Internal refactor: split the CMS experiments view module (`cms/experiments/views.py`) into a flow-focused `cms/experiments/views/` package (`_scripts`, `_experiments`, `_downloads`, `_ajax`) behind a stable `cms.experiments.views` facade. View handlers, URL routing, templates, and responses are unchanged; the split keeps each view module focused on a single request/response flow while orchestration stays in `cms.experiments.services`.
